feat: Add email input element as prop
This change adds an email input element to the LoginInput component as a
prop. It allows the Login container to pass the userEmail state and
setUserEmail function as props to the LoginInput component, enabling the
user to enter their email address. The email icon is also included in
the input field. The addition of this feature enhances the user
experience by providing a convenient and intuitive way to input email
information. This change addresses the need for user input in the login
process.